# PAM Simplification: Claude + Tools Architecture