--- a/.env.example
+++ b/.env.example
@@ -39,4 +39,8 @@ VITE_FETCH_TIMEOUT=10000
 VITE_SENTRY_DSN=https://your-sentry-dsn-here@o0000000000000000.ingest.de.sentry.io/0000000000000000
 VITE_SENTRY_ENVIRONMENT=development
 
+# Anthropic Claude API Configuration
+# API key for Claude AI integration (for SimplePAM component)
+VITE_ANTHROPIC_API_KEY=your_anthropic_api_key_here
+
 # Other environment variables can go here