--- a/src/App.tsx
+++ b/src/App.tsx
@@ -29,8 +29,9 @@ const UpdatePassword = lazy(() => import("@/pages/UpdatePassword"));
 const ThankYouDigistore24 = lazy(() => import("@/pages/ThankYouDigistore24"));
 import ProtectedRoute from './components/ProtectedRoute';
 import AdminProtection from './components/admin/AdminProtection';
-import { LazyPamIntegrationProvider } from './components/pam/LazyPamIntegrationProvider';
-import { PamProvider } from './context/PamContext';
+// OLD PAM IMPORTS - COMMENTED OUT FOR REMOVAL
+// import { LazyPamIntegrationProvider } from './components/pam/LazyPamIntegrationProvider';
+// import { PamProvider } from './context/PamContext';
 import { StagingBanner } from './components/StagingBanner';
 import { logEnvironmentInfo } from './config/environment';
 import { logEnvironmentStatus } from './config/env-validator';
@@ -48,6 +49,7 @@ const PamVoiceTest = lazy(() => import('./pages/PamVoiceTest'));
 const SiteQALog = lazy(() => import('./pages/SiteQALog'));
 const FreshTripPlannerTest = lazy(() => import('./pages/FreshTripPlannerTest'));
 const WheelsSimple = lazy(() => import('./pages/WheelsSimple'));
+const SimplePamTest = lazy(() => import('./pages/SimplePamTest'));
 
 
 const queryClient = new QueryClient();
@@ -70,8 +72,9 @@ function App() {
               <OfflineProvider>
                 <ExpensesProvider>
                   <WheelsProvider>
-                    <PamProvider>
-                      <LazyPamIntegrationProvider>
+                    {/* OLD PAM PROVIDERS - COMMENTED OUT FOR REMOVAL */}
+                    {/* <PamProvider> */}
+                      {/* <LazyPamIntegrationProvider> */}
                         <ScrollToTop />
                         <RouteMonitor />
                         <Layout>
@@ -114,13 +117,14 @@ function App() {
                                   <Route path="/qa" element={<SiteQALog />} />
                                   <Route path="/fresh-trip-planner" element={<FreshTripPlannerTest />} />
                                   <Route path="/wheels-simple" element={<WheelsSimple />} />
+                                  <Route path="/simple-pam-test" element={<SimplePamTest />} />
                                   <Route path="*" element={<div className="container p-8 text-center"><h1 className="text-2xl font-bold mb-4">404 - Page Not Found</h1><p>The page you're looking for doesn't exist.</p></div>} />
                                 </Routes>
                             </Suspense>
                           </div>
                         </Layout>
-                      </LazyPamIntegrationProvider>
-                    </PamProvider>
+                      {/* </LazyPamIntegrationProvider> */}
+                    {/* </PamProvider> */}
                   </WheelsProvider>
                 </ExpensesProvider>
               </OfflineProvider>