--- a/src/services/claude/index.ts
+++ b/src/services/claude/index.ts
@@ -1,0 +1,14 @@
+// Export main ClaudeService for easy importing
+export { 
+  ClaudeService, 
+  ClaudeServiceError, 
+  ClaudeErrorType,
+  claudeService as default 
+} from './claudeService';
+
+export type { 
+  ChatMessage, 
+  StreamingResponse, 
+  ClaudeConfig, 
+  ChatOptions 
+} from './claudeService';