… .github/workflows/main-branch-protection.yml
… ACCESSIBILITY_REPORT.md
… DAY_12_REMOVAL_PROGRESS.md
… DEPLOYMENT_CHECKLIST.md
… FINAL_TEST_REPORT.md
… INTEGRATION_TEST_SUITE_REPORT.md
… OLD_PAM_REMOVAL_PLAN.md
… PAM_IMPLEMENTATION_GUIDE.md
… PAM_SIMPLIFICATION_COMPREHENSIVE_PLAN.md
… PERFORMANCE_REPORT.md
… PUSH_TO_MAIN.md
… STAGING_CORS_ISSUE_REPORT.md
… TEST_INTEGRATION_SUMMARY.md
… docs/ROLLBACK_PLAN.md
… docs/USER_ANNOUNCEMENT.md
… netlify.toml
… src/components/pam/SimplePAM.css
… src/components/pam/SimplePAM.integration.test.tsx
… src/components/pam/SimplePAM.test.tsx
… src/components/pam/SimplePAM.tsx
… src/components/pam/SimplePAM.voice.test.tsx
… src/components/pam/accessibility/AccessibilityTestSuite.tsx
… src/components/pam/accessibility/AccessiblePAM.tsx
… src/components/pam/accessibility/AccessibleVirtualizedConversation.tsx
… src/components/pam/analytics/AnalyticsDashboard.tsx
… src/components/pam/performance/LazyVoiceComponents.tsx
… src/components/pam/performance/OptimizedPamComponents.tsx
… src/components/pam/performance/VirtualizedConversation.tsx
… src/components/pam/voice/MOBILE_INTEGRATION_GUIDE.md
… src/components/pam/voice/MobileVoiceTestChecklist.md